requirements.txt
streamlit==0.91.0
requests==2.26.0
Pillow==8.3.1
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.6.3